backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
openai==1.12.0
networkx==3.2.1
